Added a function 'run_in_thread()' to run functions in a thread
This function takes two arguments. Both of the arguments are functions,
the first argument is the 'slow function' that has to be run in a
separate thread. The second function is a callback that executes when
the first one returns.

There are some severe limitations to the current approach.
 - It is only allowed to pass one argument
 - We have to manually keep track of and clean up the threads
The cleaning is done by setting a timer that cleans the list of threads
every second.

Currently the threads method is only implemented to run
'self.instrument.spectrom.get_corrected_spectra()' in a thread in debug
mode. It should be used for all functions that are slow and blocking.