Made a new class called SensorStateUpdateManager
Also deleted the old STSVIS object. @lisapro and I wrote this code
together. The idea is to make the main panel code a little smaller and
to centralize state management to a certain extent.

The class should control reading values from sensors and if new values
have been fetched, it is responsible for updating the relevant Widgets
in the main panel.

In addition an object of this class can control if the updating is done
via a live loop or if the live loop is disabled and the updates have to
be called manually.