Consistently implemented self.major_mode
The main idea is that changes to enabled/disabled of buttons is governed
by the two methods that change self.major_mode. So on every state change
in set_major_mode or unset_major_mode all the correct buttons and
widgets are enabled or disabled.

To properly manage the modes and additional context manager was added.
It is called ongoing_major_mode_contextmanager(mode) and takes a mode
as it's argument. It ensures that set and unset are called for this mode
on entry and exit of the context.

Started using the logging library, it has yet to be integrated with the
log box in the interface.