fix autostart and shortcut for autostart on the desktop
In 16211c2638ff6c67a20b69d59ba5c7669820a5b1 the commands to change
directory were removed. This broke the autostart functionality. To keep
the local version working we can't uncomment these two lines:
# os.chdir('/home/pi/pHox')
# os.system('clear')
So instead of doing that the install3.sh script is changed. The change
is to set the working directory correctly before executing the GUI.