Config tab table changed
Added new rows to table in Config tab
Combobox for Temperature probe ids
Checkbox for temp probe "is calibrated or not"

The changes made to keep track of temperature probe installed in the
corresponding box
and to keep track if it was calibrated or not.
If not,  the default calibration coefficients are use.
The calibration coefficients are saved in evl files along with the rest
of the data.
btn manual mode disabling fixed for single sample mode