Added initial version of automatic flowcheck
Added a button to the manual button screen that runs a flow check. The
flowcheck is implemented as specified by @mar1yS. This flowcheck is
meant to check that the ferrybox system is supplying enough flow to the
phox box to make good measurements. The flowcheck works like this:

1.) Close the ferrybox valve to cut off flow
2.) Make a baseline measurement
3.) dye and stir the water
4.) Make a new measurement and compute the attenuation compared to blank
5.) Reopen the ferrybox valve to re-enable flow
6.) Make continuous measurements until too much time (20s) has passed
or until the water is clean again (20% of attenuation left)