change condition for autoadjust (maximum 15 steps not 30)
Added status finished measurement at the end
Message "wasn't able to adjust" in the status box
Check continuous button at the autostart even if pump is off.
the timer is 5 minutes, during ths 5 minutes, the checked for pump state
will be run and it will stop the cont. mode timer
But this way is more consistent, by clicking the cont. mode timer we
get all major states needed for the further state checks.