Added quit() and wait() method to thread's result_returner
It should fix QThread terminated while another thread is running problem
that was happening in all boxes after running for a long time.

quit() tells the thread's event loop to exit with return code 0(success)
wait() method blocks the thread until either of these conditions is met:
* The thread associated with this QThread object has finished execution
(i.e. when it returns from run()). This function will return true if the
 thread has finished. It also returns true if the thread has not been
started yet.
* if time parameter has elapsed (but the is no time in our call,
so it does not expire)